Move string_escaping to utils.
