+++ src/ansys/common/variableinterop/utils/string_escaping.py
